Fix bower dependency path when in child directory
I found this bug while investigating
https://github.com/emberjs/data/issues/4096

Basically, when you're in a child directory of your app, such as
`your-app/app/`, the dependency checks for bower components fail because
they cannot find the version.

This is because the version checker is using `project.bowerDirectory` to
build the path to the dependency. Since `project.bowerDirectory`
defaults to the name "bower_components" and not the actual path, this
doesn't work unless you're inside your app's root directory when the
dependency check is run.