[bugfix] Allow VersionChecker to work with Projects
VersionChecker currently can receive either an Addon or
EmberApp/EmberAddon to check dependencies against. If the user is
checking for NPM dependencies, then VersionChecker will attempt to
resolve the `package.json` of the Addon/App with a
`basedir = (app || addon).root`. Addon's know their root directory, but
EmberApp's do not - it is stored on their `project`. So, when the NPM
checker goes to resolve `package.json` of an EmberApp, it always
resolves with `undefined`, which defaults to the current NPM project.

In most cases this works just fine. However, it breaks whenever you do
`npm link`, which can cause heisenbugs while working in local projects
and generally is not fun. Additionally this same problem is causing
issues in `lerna` based projects, as they symlink the packages of the
projects together rather than install the standard way.

The ideal solution would be to change the API to only allow users to
provide an `Addon` or a `Project`, and force users to distinguish
between when they want dependencies to check of the base app and when
they want to check dependencies of a parent addon. For instance,
checking the Ember version _only_ makes sense when checking it against
the root application, so users should always do
`new VersionChecker(addon.project).forEmber()`. This would be more
verbose in some cases, but would prevent this ambiguity.

The short solution proposed by this PR moves us incrementally in that
direction by allowing the `VersionChecker` to receive a Project in the
first place, by checking the addon to see if it has a `project` field
first (as in the case of EmberApp/Addon) or if it has `root` and
`bowerDirectory` itself, throwing if neither is fulfilled.