[feat] Adds `projectNpm` as a source
Currently, when using `forEmber` we are always looking for a project
level version, but we first attempt to check the addon's own
dependencies. When linking addons in development, this breaks because
the addons will have their own version of `ember-source` installed, and
the VersionChecker will default to that version instead of resolving the
project's root version.

This problem applies to peer dependencies in general, and can cause
surprising failures, particularly when a project is switching from Bower
to NPM. Adding a `projectNpm` "location" to the version checker prevents
this ambiguity from arising in the first place since we explicitly ask
for dependencies at the project level or the local level. This also
allows us to maintain a unified, simple API for all use-cases:

```js

included(parent) {
  checker = new VersionChecker(parent);

  // Check the immediate parent's version of 'foo', always works even
  // if the immediate parent is the root project
  checker.for('foo', 'npm');

  // Check the root project's version of 'foo', no need to create a
  // second checker or use `_findHost`
  checker.for('foo', 'projectNpm');

  checker.forEmber(); // Always correct, even with linking
}
```