Support resolution with Yarn PnP.
Yarn 1.12.x supports the new PnP system to allow folks to avoid having a
`node_modules/` folder in _each_ project. To implement this, they create
a static file mapping table for each package and then require them
directly from the global Yarn cache.

Since a PnP project is no longer following the "normal" node module
resolution rules changes were needed in the `resolve` package. Those
changes landed in `resolve@1.9.0` and are leveraged by Yarn 1.13.0+.

In the mean time (throughout the Yarn 1.12.x lifespan), a simple
`resolve.sync` isn't quite good enough so we now detect if we are
running in a PnP scenario (by attempting to require the `pnpapi`
module). If we are, we use the public API provided by Yarn's `pnpapi`
module to resolve the package being checked for relative to the provided
root.