Remove bower references
Bower support [was removed](https://github.com/ember-cli/ember-cli-version-checker/pull/165) in the v5.0.0 release, but the README still had a few mentions of it. This removes those mentions. I tried to update the first example ("include different templates") to use a possibly-more-realistic version number as the pivot (3.4 is when angle bracket invocations were added).

Also fix a small typo later in the README.